Add files via upload
This repository contains the files for the Python course at El Zero Web School. These files are intended for use during the course and include exercises and examples to help solidify your understanding of Python programming. Feel free to use these files as a resource to enhance your learning experience.