Add files via upload
Python Course Tasks
This repository contains the tasks for the Python course, organized by topic. Each folder corresponds to a specific topic covered in the course, and includes a set of tasks designed to reinforce your understanding of the material. The tasks vary in difficulty and cover a range of concepts, from basic syntax to more advanced topics like object-oriented programming and data analysis.

Feel free to use these tasks to practice your Python skills and deepen your understanding of the language. Solutions to the tasks are not provided, but you're encouraged to work with fellow students or seek help from the course instructor if you need assistance. Good luck and happy coding!